+++ CalculateModel/StockFunction/HHV.cs
@@ -31,28 +31,28 @@ namespace ATrade.CalculateModel
             if (count < 1)
                 return null;
 
-            decimal hhight = 0;
+            double hhight = 0;
             object[] result=new object[data.Length];
 
             for (int i = 0; i < data.Length; i++)
             {
                 
-                if ((decimal)data[i] > hhight)
+                if ((double)data[i] > hhight)
                 {
-                    hhight = (decimal)data[i];
+                    hhight = (double)data[i];
                 }
 
                 if (i>=count)
                 {
                     int setIndex=i-count;
                     result[setIndex]=hhight;
-                    if (hhight == (decimal)data[setIndex])
+                    if (hhight == (double)data[setIndex])
                     {
                         hhight = 0;
                         for (int j = setIndex + 1; j <= i; j++)
                         {
-                            if ((decimal)data[j] > hhight)
-                                hhight = (decimal)data[j];
+                            if ((double)data[j] > hhight)
+                                hhight = (double)data[j];
                         }
                     }
                 }
@@ -61,8 +61,8 @@ namespace ATrade.CalculateModel
             hhight = 0;
             for (int i = result.Length-1; i >= result.Length-count; i--)
             {
-                if (hhight < (decimal)data[i])
-                    hhight = (decimal)data[i];
+                if (hhight < (double)data[i])
+                    hhight = (double)data[i];
                 result[i] = hhight;
             }
