+++ CalculateModel/StockFunction/Min.cs
@@ -21,13 +21,13 @@ namespace ATrade.CalculateModel
             {
                 object[] p1 = (object[])param1;
                 result.Results = new object[p1.Length];
-                decimal m = decimal.MaxValue;
+                double m = double.MaxValue;
                 if (param2 == null)
                 {
                     for (int i = p1.Length-1; i >=0; i--)
                     {
-                        if (p1[i].ToDecimal(2) < m)
-                            m = p1[i].ToDecimal(2);
+                        if (p1[i].ToDouble(2) < m)
+                            m = p1[i].ToDouble(2);
                         result.Results[i] = m;
                     }
                 }
@@ -40,15 +40,15 @@ namespace ATrade.CalculateModel
                             throw new ExpressErrorException("Min方法两边操作数不相等");
                         for (int i = 0; i < p1.Length; i++)
                         {
-                            result.Results[i] = Math.Min(p1[i].ToDecimal(2), p2[i].ToDecimal(2));
+                            result.Results[i] = Math.Min(p1[i].ToDouble(2), p2[i].ToDouble(2));
                         }
                     }
                     else
                     {
-                        decimal p2 = param2.ToDecimal(2);
+                        double p2 = param2.ToDouble(2);
                         for (int i = 0; i < p1.Length; i++)
                         {
-                            result.Results[i] = Math.Min(p1[i].ToDecimal(2), p2);
+                            result.Results[i] = Math.Min(p1[i].ToDouble(2), p2);
                         }
                     }
                 }
@@ -57,12 +57,12 @@ namespace ATrade.CalculateModel
             {
                 if (!(param2 is object[]))
                     throw new ExpressErrorException("Max方法参数错误，第二参数不是集合");
-                decimal p1 = param1.ToDecimal(2);
+                double p1 = param1.ToDouble(2);
                 object[] p2 = (object[])param2;
                 result.Results = new object[p2.Length];
                 for (int i = 0; i < p2.Length; i++)
                 {
-                    result.Results[i] = Math.Min(p1, p2[i].ToDecimal(2));
+                    result.Results[i] = Math.Min(p1, p2[i].ToDouble(2));
                 }
             }
 
@@ -73,8 +73,8 @@ namespace ATrade.CalculateModel
         {
             return new CalResult
             {
-                ResultType = typeof(decimal),
-                Result = Math.Min(decimal.Parse(param1.ToString()), decimal.Parse(param2.ToString()))
+                ResultType = typeof(double),
+                Result = Math.Min(double.Parse(param1.ToString()), double.Parse(param2.ToString()))
             };
         }
     }
